[O2 merge] gen_imaps.c: Added -black and -white options
Written by Greg Hood (I think) at the request of folks working on
APEX/DAB staining who wanted to be able to tell the difference between
really black and really really black pixels, and so needed the option to
be able to prevent contrast adjustment that was aggressive enough to
assign all the black stuff to a value of 0. Aaron knows more about this
I think.